requirements.txt
streamlit==1.1.0
Pillow==8.0.1
opencv-python==4.4.0.46
numpy==1.19.4
